Struct cactusref::Weak

pub struct Weak<T: ?Sized> { /* fields omitted */ }

Weak is a version of Rc that holds a non-owning reference to the managed value. The value is accessed by calling upgrade on the Weak pointer, which returns an Option<Rc<T>>.

Since a Weak reference does not count towards ownership, it will not prevent the inner value from being dropped, and Weak itself makes no guarantees about the value still being present and may return None when upgraded.

A Weak pointer is useful for keeping a temporary reference to the value within Rc without extending its lifetime. It is also used to prevent circular references between Rc pointers, since mutual owning references would never allow either Rc to be dropped. For example, a tree could have strong Rc pointers from parent nodes to children, and Weak pointers from children back to their parents.

The typical way to obtain a Weak pointer is to call Rc::downgrade.

Methods

impl<T> Weak<T>

pub fn new() -> Self

Constructs a new Weak<T>, without allocating any memory.

Calling upgrade on the return value always gives None.

Examples

use cactusref::Weak;

let empty: Weak<i64> = Weak::new();
assert!(empty.upgrade().is_none());

pub fn as_raw(&self) -> *const T

Returns a raw pointer to the object T pointed to by this Weak<T>.

It is up to the caller to ensure that the object is still alive when accessing it through the pointer.

The pointer may be null or be dangling in case the object has already been destroyed.

Examples

use cactusref::Rc;
use std::ptr;

let strong = Rc::new("hello".to_owned());
let weak = Rc::downgrade(&strong);
// Both point to the same object
assert!(ptr::eq(&*strong, weak.as_raw()));
// The strong here keeps it alive, so we can still access the object.
assert_eq!("hello", unsafe { &*weak.as_raw() });

drop(strong);
// But not any more. We can do weak.as_raw(), but accessing the pointer
// would lead to undefined behaviour.
// assert_eq!("hello", unsafe { &*weak.as_raw() });

pub fn into_raw(self) -> *const T

Consumes the Weak<T> and turns it into a raw pointer.

This converts the weak pointer into a raw pointer, preserving the original weak count. It can be turned back into the Weak<T> with from_raw.

The same restrictions of accessing the target of the pointer as with as_raw apply.

Examples

use cactusref::{Rc, Weak};

let strong = Rc::new("hello".to_owned());
let weak = Rc::downgrade(&strong);
let raw = weak.into_raw();

assert_eq!(1, Rc::weak_count(&strong));
assert_eq!("hello", unsafe { &*raw });

drop(unsafe { Weak::from_raw(raw) });
assert_eq!(0, Rc::weak_count(&strong));

pub unsafe fn from_raw(ptr: *const T) -> Self

Converts a raw pointer previously created by into_raw back into Weak<T>.

This can be used to safely get a strong reference (by calling upgrade later) or to deallocate the weak count by dropping the Weak<T>.

It takes ownership of one weak count. In case a null is passed, a dangling Weak is returned.

Safety

The pointer must represent one valid weak count. In other words, it must point to T which is or was managed by an Rc and the weak count of that Rc must not have reached 0. It is allowed for the strong count to be 0.

Examples

use cactusref::{Rc, Weak};

let strong = Rc::new("hello".to_owned());

let raw_1 = Rc::downgrade(&strong).into_raw();
let raw_2 = Rc::downgrade(&strong).into_raw();

assert_eq!(2, Rc::weak_count(&strong));

assert_eq!("hello", &*unsafe { Weak::from_raw(raw_1) }.upgrade().unwrap());
assert_eq!(1, Rc::weak_count(&strong));

drop(strong);

// Decrement the last weak count.
assert!(unsafe { Weak::from_raw(raw_2) }.upgrade().is_none());

impl<T: ?Sized> Weak<T>

pub fn upgrade(&self) -> Option<Rc<T>>

Attempts to upgrade the Weak pointer to an Rc, extending the lifetime of the value if successful.

Returns None if the value has since been dropped.

Examples

use cactusref::Rc;

let five = Rc::new(5);

let weak_five = Rc::downgrade(&five);

let strong_five: Option<Rc<_>> = weak_five.upgrade();
assert!(strong_five.is_some());

// Destroy all strong pointers.
drop(strong_five);
drop(five);

assert!(weak_five.upgrade().is_none());

pub fn strong_count(&self) -> usize

Gets the number of strong (Rc) pointers pointing to this value.

If self was created using Weak::new, this will return 0.

pub fn weak_count(&self) -> Option<usize>

Gets the number of Weak pointers pointing to this value.

If self was created using Weak::new, this will return None. If not, the returned value is at least 1, since self still points to the value.

pub fn ptr_eq(&self, other: &Self) -> bool

Returns true if the two Weaks point to the same value (not just values that compare as equal).

Notes

Since this compares pointers it means that Weak::new() will equal each other, even though they don't point to any value.

Examples

use cactusref::Rc;

let first_rc = Rc::new(5);
let first = Rc::downgrade(&first_rc);
let second = Rc::downgrade(&first_rc);

assert!(first.ptr_eq(&second));

let third_rc = Rc::new(5);
let third = Rc::downgrade(&third_rc);

assert!(!first.ptr_eq(&third));

Comparing Weak::new.

use cactusref::{Rc, Weak};

let first = Weak::new();
let second = Weak::new();
assert!(first.ptr_eq(&second));

let third_rc = Rc::new(());
let third = Rc::downgrade(&third_rc);
assert!(!first.ptr_eq(&third));