Trait cactusref::Adoptable

pub unsafe trait Adoptable {
    fn adopt(this: &Self, other: &Self);

    fn unadopt(this: &Self, other: &Self);
}

Perform bookkeeping to link two objects with an owned reference.

Calling Adoptable::adopt builds an object graph which can be used by implementors to detect cycles.

Warning: this trait is unsafe because if it is implemented incorrectly, memory may leak or be double freed.

Required methods

fn adopt(this: &Self, other: &Self)

Perform bookkeeping to record that this has an owned reference to other. Adoption is a one-way link.

fn unadopt(this: &Self, other: &Self)

Perform bookkeeping to record that this no longer has an owned reference to other. Adoption is a one-way link.

Loading content...