Crate cactusref

Structs

Rc
A single-threaded reference-counting pointer. 'Rc' stands for 'Reference Counted'.

Weak
Weak is a version of Rc that holds a non-owning reference to the managed value. The value is accessed by calling upgrade on the Weak pointer, which returns an Option<Rc<T>>.

Traits

Adoptable
Take strong ownership of an object without forming a reference cycle.

Reachable
Reachability tests for objects wrapped by Rc. This trait is unsafe.

Type Definitions

CactusRef
Cactus alias for Rc.

CactusWeakRef
Cactus alias for Weak.