Struct cactusref::Rc

pub struct Rc<T: ?Sized> { /* fields omitted */ }

A single-threaded reference-counting pointer. 'Rc' stands for 'Reference Counted'.

This Rc differs from the Rc in std by adding support for detecting and deallocating orphaned cycles of references. An orphaned cycle is one in which all objects are only owned by other members of the cycle.

See the module-level documentation for more details.

The inherent methods of Rc are all associated functions, which means that you have to call them as e.g., Rc::get_mut(&mut value) instead of value.get_mut(). This avoids conflicts with methods of the inner type T.

Methods

impl<T> Rc<T>

pub fn new(value: T) -> Self

Constructs a new Rc<T>.

Examples

use cactusref::Rc;

let five = Rc::new(5);

pub fn pin(value: T) -> Pin<Self>

Constructs a new Pin<Rc<T>>. If T does not implement Unpin, then value will be pinned in memory and unable to be moved.

pub fn try_unwrap(this: Self) -> Result<T, Self>

Returns the contained value, if the Rc has exactly one strong reference.

Otherwise, an Err is returned with the same Rc that was passed in.

This will succeed even if there are outstanding weak references.

Examples

use cactusref::Rc;

let x = Rc::new(3);
assert_eq!(Rc::try_unwrap(x), Ok(3));

let x = Rc::new(4);
let _y = Rc::clone(&x);
assert_eq!(*Rc::try_unwrap(x).unwrap_err(), 4);

impl<T: ?Sized> Rc<T>

pub fn into_raw(this: Self) -> *const T

Consumes the Rc, returning the wrapped pointer.

To avoid a memory leak the pointer must be converted back to an Rc using Rc::from_raw.

Examples

use cactusref::Rc;

let x = Rc::new("hello".to_owned());
let x_ptr = Rc::into_raw(x);
assert_eq!(unsafe { &*x_ptr }, "hello");

pub unsafe fn from_raw(ptr: *const T) -> Self

Constructs an Rc from a raw pointer.

The raw pointer must have been previously returned by a call to a Rc::into_raw.

This function is unsafe because improper use may lead to memory problems. For example, a double-free may occur if the function is called twice on the same raw pointer.

Examples

use cactusref::Rc;

let x = Rc::new("hello".to_owned());
let x_ptr = Rc::into_raw(x);

unsafe {
    // Convert back to an `Rc` to prevent leak.
    let x = Rc::from_raw(x_ptr);
    assert_eq!(&*x, "hello");

    // Further calls to `Self::from_raw(x_ptr)` would be memory unsafe.
}

// The memory was freed when `x` went out of scope above, so `x_ptr` is now dangling!

pub fn into_raw_non_null(this: Self) -> NonNull<T>

Consumes the Rc, returning the wrapped pointer as NonNull<T>.

Examples

use cactusref::Rc;

let x = Rc::new("hello".to_owned());
let ptr = Rc::into_raw_non_null(x);
let deref = unsafe { ptr.as_ref() };
assert_eq!(deref, "hello");

pub fn downgrade(this: &Self) -> Weak<T>

Creates a new Weak pointer to this value.

Examples

use cactusref::Rc;

let five = Rc::new(5);

let weak_five = Rc::downgrade(&five);

pub fn weak_count(this: &Self) -> usize

Gets the number of Weak pointers to this value.

Examples

use cactusref::Rc;

let five = Rc::new(5);
let _weak_five = Rc::downgrade(&five);

assert_eq!(1, Rc::weak_count(&five));

pub fn strong_count(this: &Self) -> usize

Gets the number of strong (Rc) pointers to this value.

Examples

use cactusref::Rc;

let five = Rc::new(5);
let _also_five = Rc::clone(&five);

assert_eq!(2, Rc::strong_count(&five));

pub fn get_mut(this: &mut Self) -> Option<&mut T>

Returns a mutable reference to the inner value, if there are no other Rc or Weak pointers to the same value.

Returns None otherwise, because it is not safe to mutate a shared value.

See also make_mut, which will clone the inner value when it's shared.

Examples

use cactusref::Rc;

let mut x = Rc::new(3);
*Rc::get_mut(&mut x).unwrap() = 4;
assert_eq!(*x, 4);

let _y = Rc::clone(&x);
assert!(Rc::get_mut(&mut x).is_none());

pub fn ptr_eq(this: &Self, other: &Self) -> bool

Returns true if the two Rcs point to the same value (not just values that compare as equal).

Examples

use cactusref::Rc;

let five = Rc::new(5);
let same_five = Rc::clone(&five);
let other_five = Rc::new(5);

assert!(Rc::ptr_eq(&five, &same_five));
assert!(!Rc::ptr_eq(&five, &other_five));

impl<T: Clone> Rc<T>

pub fn make_mut(this: &mut Self) -> &mut T

Makes a mutable reference into the given Rc.

If there are other Rc pointers to the same value, then make_mut will clone the inner value to ensure unique ownership. This is also referred to as clone-on-write.

If there are no other Rc pointers to this value, then Weak pointers to this value will be dissassociated.

See also get_mut, which will fail rather than cloning.

Examples

use cactusref::Rc;

let mut data = Rc::new(5);

*Rc::make_mut(&mut data) += 1;        // Won't clone anything
let mut other_data = Rc::clone(&data);    // Won't clone inner data
*Rc::make_mut(&mut data) += 1;        // Clones inner data
*Rc::make_mut(&mut data) += 1;        // Won't clone anything
*Rc::make_mut(&mut other_data) *= 2;  // Won't clone anything

// Now `data` and `other_data` point to different values.
assert_eq!(*data, 8);
assert_eq!(*other_data, 12);

Weak pointers will be dissassociated:

use cactusref::Rc;

let mut data = Rc::new(75);
let weak = Rc::downgrade(&data);

assert!(75 == *data);
assert!(75 == *weak.upgrade().unwrap());

*Rc::make_mut(&mut data) += 1;

assert!(76 == *data);
assert!(weak.upgrade().is_none());